Updations in JS Date formatter

+++ ImportantConcepts/DateConversion.js
@@ -1,10 +1,10 @@
 // ******Date formatter using new Date********
 
-// Format current date to specified format -
+// Format current date to specified format - Feb 24, 2024
 
 const formattedDate = new Date().toLocaleDateString("en-US", {
   year: "numeric",
-  month: "long",
+  month: "short",
   day: "numeric",
 });
